Update patch set 1
Patch Set 1:

Welp, after reading the docs for glfwWaitEvents and glfwPostEmptyEvent more closely, I realize that this isn't just a temporary workaround, but in fact it's the right thing to do. From glfwWaitEvents's docs:

> If no windows exist, [glfwWaitEvents] returns immediately. For synchronization of threads in applications that do not create windows, use your threading library of choice.

The bug in GLFW is that it shouldn't be crashing, but we shouldn't be trying to use glfwWaitEvents before the first window is created either.

I'll update this CL accordingly before submitting.

Patch-set: 1
Reviewer: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>