shiny/driver/mtldriver: select on receiving from newWindowCh
This is a fixup to CL 208877. The intention there was to select
between receiving a value from the newWindowCh or done channel.
Instead, it was selecting between sending to newWindowCh and
receiving from done. That means a shiny program that doesn't
open a window could never exit. This change fixes that.

Change-Id: I1c09ea5f7bec6c9714f20428fdd90d1b756f515d