Update patch set 8
Patch Set 8:

I am not familiar with the exact changes that have been introduced in the last few patchsets, but it looks like -base=none is not acting as expected:

 panic: could not parse base version: version "none" does not start with 'v'`

I would also like to comment that the removal of auto-detecting the latest version, makes ci integration difficult. Would it be possible to add this functionality back via -base=auto, possibly as the default?

Patch-set: 8
CC: Gerrit User 31130 <31130@62eb7196-b449-3ce5-99f1-c037f21e1705>