Update patch set 9
Patch Set 9: Run-TryBot+1

(8 comments)

> I am not familiar with the exact changes that have been introduced in the last few patchsets, but it looks like -base=none is not acting as expected:
> 
>  panic: could not parse base version: version "none" does not start with 'v'`

Fixed, thanks for pointing that out.

> I would also like to comment that the removal of auto-detecting the latest version, makes ci integration difficult. Would it be possible to add this functionality back via -base=auto, possibly as the default?

Yes, but not in this CL. Bryan had the same comment. It's now at the top of the TODO list.

Previously, gorelease loaded the base version from the local VCS repo. It could automatically find the previous version that was an ancestor of the current commit. Now, gorelease loads the base version with 'go mod download'. The previous version is not well defined, but we could do something intelligent with 'go list -m -versions' and the -version flag (if specified).

Patch-set: 9
Reviewer: Gerrit User 17092 <17092@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1