cmd/gorelease: infer base version if unspecified
If the -base flag is not set on the command line, it is now inferred
from the -version flag and available versions. If the -version flag is
given, the base version will be the highest available release version
lower than -version. Otherwise, the base version will be the highest
release version. Pre-release versions are not considered.

If there are no appropriate versions, and the release version is
unspecified or appears to be the first release (e.g., v0.1.0, v2.0.0),
the inferred base version will be "none" (meaning no comparison will
be made). Otherwise, an error will be shown.

Updates golang/go#26420

Change-Id: Iee45a4183f3e4a219c02a69b5d16a3cc5478644c