cmd/gorelease: sort packages returned by packages.Load.
packages.Load makes no guarantee about the order in which packages are
returned. Previously, we assumed that when an explicit list of
packages are passed in (no patterns), packages are returned in the
same order. packages.Load uses 'go list -deps' internally, which
prints packages in depth-first post-order.

Fixes golang/go#37756

Change-Id: I1999815e778cd2a28145027c9dba19bed51c25f2