cmd/gorelease: extract loading functions out of makeReleaseReport
This is a refactoring CL: there should be no visible change in behavior.

There are several feature requests related to loading the base and
release versions from different places or with different module
paths. This CL enables that work by extracting two functions out of
makeReleaseReport.

loadLocalModule loads the release version from a local directory.

loadDownloadedModule downloads the base version and loads it from the
module cache.

In the future, base and release may not necessarily be remote and local.

For golang/go#39666

Change-Id: I37955caacf02479b7a3eaf03782f140a42f3a87e