x/exp/cmd/gorelease: avoid using -mod=mod and support lazy module loading
This causes us to have to interpret some `go list` errors, which is unfortunate.
If any reviewer sees a way to avoid this, lets me know.

This causes us to have to exit earlier than before when we don't have a go.sum,
or have an empty go.sum. Not sure if that's unfortunate or not. If any reviewer
sees a way to avoid that, please let me know.

TODO: Is there a way to avoid the three strings.Contains we're doing here?

Fixes golang/go#41456

Change-Id: I66f12fe0325e4fed432d8b10c1fe977068e4a030