shiny/iconvg: add a decode + encode round-trip test
Given we promise deterministic encoding, and decoding is also
(trivially) deterministic, it's viable to add another test to
ensure that decoding and encoding an IconVG file gives us the
same bytes back.

Since the Encoder.HighResolutionCoordinates option is not a
part of the image metadata but needed for this test to work,
obtain it from the filename.

Also remove some comparisons of adj that can never be false
because no value of type uint8 is less than 0.

Change-Id: I356ec01c8dd582aaeff14b39bb0272ca93eb1d2a