Added a needed feature to winfsnotify
The package lacks the ability to watch an entire tree for changes.  To address that, I added a Boolean to the system, and the AddDeepWatch() and DeepWatch() functions to enable the 'bWatchSubtree' parameter of ReadDirectoryChanges() with this Boolean.  This will allow the watch to report events anywhere within the watched folder tree.

The original functions, AddWatch() and Watch(), continue to behave shallow as they did before--only watching the specified folder--so existing code will not be impacted.

Change-Id: I2d15ba28391c6e8638cebea15218b29b965a0fb3
GitHub-Last-Rev: 1305fe561f77c056d8f2d30112c0913c4f350430
GitHub-Pull-Request: golang/exp#16