event: simplify label slice storage optimization
Now that Builders are on the heap, we can put the static storage
for labels in the Builder, and use it for Event.Labels.

This simplifies dealing with the static label optimization: now there
is only one slice of labels, and its backing array requires no
additional allocation for small sizes.

Change-Id: Ie6badf70b9666aa23d6ab0a287a831d69feb1249