event: suggestion: event.Start returns end
The majority of calls to start a trace span will be followed by a deferred call to end one:

    ctx = event.To(ctx).Start(name)
    defer event.To(ctx).End()

The main problem with this is that it is easy to forget that end must
be called. It's also more verbose than necessary.

Instead, have event.Start return both a context and a function that
can be deferred:

    ctx, end := event.To(ctx).Start(name)
    defer end()

The second return value will make it hard to forget the defer, and it
follows a common Go pattern.

You can also ignore it and call event.To(ctx).End() elsewhere if that
suits you.

Lastly, remove Builder.Start. It invites misuse by end users: they
should always do tracing with event.Start to properly construct the
span tree. (If you really want to do it, you can still call Deliver.)

Change-Id: Ibe7780ce5277e9d6dac5406e879a35d727a89eff