cmd/gorelease: fix test runner, and implementation of existing versions check
Currently, all tests pass because the test runner uses the for-loop-scoped
"test" in multiple parallel tests. This change locally scopes it to the for
block so that each parallel test run has its own "test".

That causes some tests to fail. The primary failure is that existing version
check is happening in loadLocalModule. loadLocalModule uses the release version
(the -version flag), whereas loadDownloadedModule uses the base version. Since
loadLocalModule uses the release version, prepareLoadDir (performed inside both
load* functions) will use the release version for the replace directive, which
means "go list" will report the release version. The existing version check uses
"go list", so it's guaranteed to always fail: the release specified will always
be reported as an existing release!

The change is to move the existing version check to loadDownloadedModule, which
performs "go list" as before but now it's only checking things which already
exist, not which the user is proposing should exist.

Change-Id: I5a9776d6ae1af6718b3c39946945c4bbb108a2fc