event/otel: metrics
This CL shows how a metric handler could be constructed to use event
metrics with Open Telemetry.

The first time it sees an event.Metric, the handler creates a matching
otel instrument and caches it. On each call, it uses the instrument to
record the metric value.

Handling of units, distributions, etc. is not required in this simple
case.  It is TBD for more complex cases.

Change-Id: I07d6f40601c7d2a801ed9fbe3cf7c24d5698f3f1