event/otel: metrics
This CL shows how a metric handler could be constructed to use event
metrics with Open Telemetry.

The handler's RegisterXXX functions associate otel metrics with event
namespace/names for metrics. When a value is recorded, the associated
metric is looked up and its Add or Record method is
called. Unregistered names are ignored.

Change-Id: I07d6f40601c7d2a801ed9fbe3cf7c24d5698f3f1