event: better metrics support
The current design of metric events is too underspecified to be
useful. A metric handler needs to know the name and value of the
metric. If all it gets is an undifferentiated list of labels, it can't
know which labels identify the metric and its value and which merely
categorize it.

This CL makes sure the metric information is in a well-specified
place.  It replaces Builder.Metric with two new methods, one for int
metrics and one for float metrics, both of which place the metric
label at the end of the list.

This CL also tackles the problem that metric names form a global
namespace. To avoid clashes between common metric names like
"request_count", the two metric methods require a namespace argument,
which could be an import path. The namespace is stored in
Event.Message.

Change-Id: I9a1a16fe83b5a646c36ca03596603ee98b71f20c