event: better metrics support
The current design of metric events is too underspecified to be
useful. A metric handler needs to know the name and value of the
metric. If all it gets is an undifferentiated list of labels, it can't
know which labels identify the metric and its value and which merely
categorize it.

One design path begins with a convention for label names. The handler
can look for these names to distinguish the metric from other labels.

This CL takes that one step further, to avoid having the handler
search the list of labels.  It replaces Builder.Metric with two new
methods, one for int metrics and one for float metrics, both of which
place the metric labels at the end of the list.

This CL also tackles the problem that metric names form a global
namespace. To avoid clashes between common metric names like
"request_count", the two metric methods require a namespace argument,
which could be an import path.

Change-Id: I9a1a16fe83b5a646c36ca03596603ee98b71f20c