racy: add demo racy program
This is currently in golang.org/x/blog/support/racy,
for use in https://blog.golang.org/race-detector,
but we really should not be encouraging use of the blog
repo by anyone other than blog authors. x/exp is a better
landing place for this.

Made it a submodule to keep it separate from any other
dependencies in this repo.

Change-Id: I109d10870b05c5547d83bb1230694be129f8e382