event: WIP: another try at metrics
NOTE: TESTS DON'T PASS

For discussion only.

The code that is recording the metric is the best place to describe
that metric. For example, an RPC layer knows that number of sent RPCs
is a counter, and that bytes received is a distribution.

We want to keep that information separate from the actually metric
recoding, both to avoid duplication and to amortize any costs
involved.

The idea in this CL is to have a lightweight representation for
metrics. The world seems to have converged on just three kinds:
counter, gauge and distribution in the terminology of this CL.

A package must first create the metrics it wants to record, by calling
NewMetric with the metric kind, a name, and a namespace. By forcing
the namespace we can prevent ambiguous metric names, such has having
many packages use the name "call_count".

It's up to the package whether it creates the metrics globally and
unconditionally (perhaps using its import path as namespace), or each
time a value of a certain type is created. A cache that has just one
global instance would use the former, while and RPC package with a
Client type would ask the caller of NewClient to pass a namespace, and
store the created metrics in the Client.

Builder.Metric records a metric value as before. Its first argument is
one of the previously created metrics, from which it can extract
whatever values it needs to build an Event. Its second argument is an
event.Value, allowing any type to be a metric value.

Change-Id: I26d0ac095fa2cd8b19201edc47350704636de89c