event: Builder.Start returns a builder
Currently, Builder.Start's second return value is a function which,
when called, sends an End event with a copy of the Start event's
builder. This makes the common case

    ctx, end := event.To(ctx).Start(name)
    defer end()

easy to write.

But it's not possible to add additional labels that way; instead a
whole new builder needs to be constructed from scratch.

So in this CL, Start's second return value is a Builder. The simple
case is almost as short:

    ctx, b := event.To(ctx).Start(name)
    defer b.End()

but now it's equally easy to add more information to the end event:

    ctx, b := event.To(ctx).Start(name)
    defer b.With(Result.Of(err)).End()

Change-Id: Ib489a2704ac78479317eeb3f1f3de0f1b38d610c