event: a new mechanism for caller detection
This avoids the need to know the stack depth to the caller by instead scanning the
stack skipping over entries that have been registered as helpers.
It also allows registration of helper packages, where every function from that package
is considered a helper function. This is used to elimitate all functions from the 
event package.

Change-Id: Id747e1a4a5ccbfa84dfd372cc33471490aa426ef