vulndb/govulncheck/binscan: more exhaustive PCLN search for ELF
Some compliation/linking/stripping strategies appear to result in the
PCLN table ending up in a .data.rel.ro section, but without a header
for the 'gopclntab' part of the section. This change implements a more
exhaustive search for the PCLN table by looking for the pclntab magic
and the go magic in the .data.rel.ro section.

Change-Id: Ibb907c80b1fe3b8bf21aa3f83adae3c65bcdf1a4