exp/vulndb/internal/audit: optimize db loading
For large projects, such as k8s, the number of modules different in
memory but with the same import path can be large. This can result in
many db queries for the same import path, which can be a problem if the
db is remote and no caching is turned on. This CL optimizes the loading
by introducing a simple in-memory caching during the loading.

The caching introduced in this CL is different from the caching of the
db client as it only avoids loading of things at the same import path.

Change-Id: I63f09382e318fa33f7e9280c8cc1ddb5694983b7