exp/vulndb/govulncheck: optimize db loading
For large projects, such as k8s, the number of modules different in
memory but with the same import path can be large. This can result in
many db queries for the same import path, which can be a problem if the
db is remote and no caching is turned on. This CL optimizes the loading
by making sure module extraction returns modules unique up to import
paths.

This CL also updates the dependencies on vulndb library.

Change-Id: I63f09382e318fa33f7e9280c8cc1ddb5694983b7