vulndb/internal/audit: filter out vulns for modules with "" version
When a module version is unknown (""), the current implementation
assumes that any vulnerability version range applies to it. This can
lead to false alarms, the most prominent example being when audit is
run on a top-level module (which will have "" version) that has known
vulnerabilities. This CL makes sure no vulnerabilities apply for a
module with an unavailable version.

Fixes golang/go#48079

Change-Id: Idd9f080f9037d105d86311b62de77f29ef4664a2