vulndb/internal/audit: filter affected fields in vulnerabilities
Filtering of vulnerabilities did not filter affected fields. This can
lead to false positives. For instance, filtering would deem a
vulnerability applicable to the current user code if a single affected
field applies to it. Other affected fields would be kept too even though
they might not apply to the code. If the code uses packages and symbols
related to that not-applicable affected fields, audit will report them.
The most prominent example of this is YAML vulnerability. This CL
filters affected fields as well.

Change-Id: I92a521a7eeadf12376a3c42f4fe69769ee1c3637