shiny/driver/gldriver: explicitly include needed X11 headers
x11.c refers to several types defined in Xlib headers, but did not
include those headers explicitly. As far as I can tell, it assumed
that they are indirectly included via egl.h — but that assumption
empirically does not always hold.

This fixes a build error on my machine. I'm not sure why it hasn't
shown up on the Go dashboard builders.

Change-Id: I70146c68ff7e4d32af63a5aa8d7a8e25445091c5