vulncheck: add forward slicing logic
This code is copied from vulndb/internal/audit.

The logic is used to compute a forward reachable set of functions
starting from program entry points. This set is effectively a
conservative estimate of the program subset relevant for vulncheck,
and is passed to vta call graph analysis (comes in future CLs).

Change-Id: I83c213c137976d5a3877eb478338affb9c72960b