vulncheck: fix support for packages whose every symbol is vulnerable
Some vulnerabilities have an empty slice for the osv.Entry Symbols
field, meaning every symbol is vulnerable. This CL adds support for such
vulnerabilities by doing the following, when an import of the
corresponding packages is seen, we list every top-level method and
function of that package and add it as a Vuln in vulncheck.Result.

Change-Id: I0edac51f7e3923ecfd3f203db80e6f7d22272dd2