slices: initial implementation of sorting functions
Implements golang/go#47619 in the exp/slices package as a
testing ground prior to inclusion in the standard library.

Relies on the modified sorting function code generator proposed
in https://go-review.googlesource.com/c/go/+/353069 to
automatically generate the code of the sorting functions.

Benchmarks comparing sort.Ints with the generic Sort function
added in this CL to sort a slice of int:

    goos: linux
    goarch: amd64
    pkg: golang.org/x/exp/slices
    cpu: Intel(R) Core(TM) i7-4771 CPU @ 3.50GHz
    BenchmarkSortInts-8          98	  12100642 ns/op
    BenchmarkSlicesSort-8        186	   6594885 ns/op

Comparing sort.Sort with SortFunc to sort a slice of struct
pointers based on one field in the struct:

    BenchmarkSortStructs-8       61	  18640003 ns/op
    BenchmarkSortFuncStructs-8   63	  16451417 ns/op

Change-Id: Ic301aae7e5b8f99144e39b8a77fde897779588ed