shiny: add go.mod file
The shiny package is not related to other exp packages.
Put it in a separate module.

This package uses a modified copy of github.com/BurntSushi/xgb.
Move that out of the vendor directory, so that current go commands work.
Use a replace directive to point to the modified copy.

Change-Id: Ib3bad0823d2651ff90a9651e405ed7bcd9de94b6