vulncheck: remove isLocal check from fetchVulnerabilities
isLocal check was added as a fix for golang/go#48079.
With the check, fetchVulnerabilities excludes vulnerabilities
for modules if their source files are outside the module cache.
The location of the module cache was determined by querying
GOMODCACHE and GOPATH environment variables of the govulncheck
process. That worked well for govulncheck when it is used
for source scanning.

The logic was copied to vulncheck API internal. However,
relying on process's GOMODCACHE/GOPATH environment variables
limit the API's utility. For example, Gopls may use different
GOMODCACHE/GOPATH for each workspace it's processing and they
can be different from the Gopls's own GOMODCACHE/GOPATH env vars.

There was an escape flag to skip this check to work with
the test environment where the module cache and GOPATH
are different from the test process's. But that is unexported;
external packages cannot utilize it and that prevents
writing tests from external packages.

This CL proposes to remove the isLocal check. Users
and applications that need to exclude vulnerabilities of
local modules, may utilize golang.org/x/vuln/client.Client
and implement filtering from GetByModule.

Change-Id: Iced93351b91a00fdc623a6d1c3076da86fbe2c70