vulncheck: fix buildSSA's FileSet handling
The FileSet used when building a new ssa program should be
the same FileSet used when building the Package input
because later Position computation is done using the SSA
program's FileSet.

Moreover, enforce this from the Source API level.
If Packages passed to Source were built with different
FileSets, that is an error.

Change-Id: I5336fcc12b6fe951e625ec1c2223439efbeccf2b