vulncheck: allow to build with go1.17 or older
vulncheck/internal/binscan uses debug/buildinfo that was added
in go1.18. That prevented use of vulncheck package completely
when building with go1.17 or older.
Stub binscan.ExtractPackagesAndSymbols depending on the version
of Go, so the part of vulncheck API that does not depend on it
can still be available for use with older versions of Go.

As a result, vulncheck.Binary will always return an error
if it is built with go1.17 or older.

Fixes golang/go#51687

Change-Id: I48d83aa227a3eabb9fbbb384d3a09c28f33ff24f