proposal: a convenience logging package
Add a log package that accommodates both structured and printf-style logging.
I've been using something like this in golang.org/x/vulndb/internal/worker
with success.

event.Log is still the low-level, zero-allocation way to produce log events,
for those who need the performance. But this new package will work for
everyone else.

In keeping with event.Log being a low-level API, remove event.Logf; its job
is taken by this new package.

Also, remove event.Error, because it's a convenience method and the
idea is that convenience is the job of this new package. (The new
package could add WithError function/methods to make up for the loss.)

Change-Id: Ib70fae587a28fe2f612749fe4d2b6fed6d5ff9eb