slog: Attrs
Define Attr, an alloc-efficient key-value pair.

Attr is like any, but does not allocate for most basic types.
It can be used for low-allocation logging.

The unsafe version, which is the default, exploits the representation
of Go strings to save space.

The ideas here are largely taken from Ian Cottrell's Label implementation,
https://cs.opensource.google/go/x/exp/+/334a2380:event/label.go.

Change-Id: I8283fefc208ab6d306d5cc32bc0a0ccc295fab57