slog: EXPERIMENT: replace Any(any) with Any[T](T)
Remove the Any function, which constructs an Attr from a value of type any,
and replace it with A[T].

Remove the Float64 and Bool constructors.

For recognized types, like float64, the generic version is faster and
does not allocate. It is slower (by about 2x) when T = any.  It is
also much slower than some type-specific constructors like String and
Int.

At present, it would allow us to remove some type-specific
constructors that would be rarely used, like Float64 and Bool.  On the
other hand, we still need accessors for those types, and they have
their own Kinds.

If the compiler implemented switching on generic parameters
(https://go.dev/issue/45380), then A[T] should match the performance
of the type-specific constructors, and we could get rid of all of
them.

Change-Id: Ia391d35b064f7116cbb40787c010dbf7d575965b