slog: benchmarks
Add benchmarks to measure the performance of LogAttrs and compare it
with Zap.

The run_benchmarks.sh script will run the benchmarks 5 times both with
and without the code to fetch the pc, and write the results to
files. That code is currently slow (200ns) so we optimistically
disable it.

Put the equivalent Zap benchmarks in a subdirectory, zapbenchmarks,
which is in a separate module to avoid having golang.org/x/exp depend
on go.uber.org/zap.  The Zap benchmarks are named identically to the
slog ones so benchstat can compare them. The
zapbenchmarks/run_benchmarks.sh script runs the benchmarks, puts the
output in a file, and rename the "pkg:" header in the file to match
the slog benchmarks package so that benchstat will compare them.

Change-Id: I88cf12f6e3b2a0eab894664cb106f3b4ec3996c4