slog: redo Record API
This CL presents a new Record API and implementation.

The API is more limited than before, to increase implementation
freedom. Attrs are added to Records in chunks instead of one at a
time, and random access is replaced by forward iteration.

The implementation is simplified as well. In addition to the inline
array, there is a linked list of slices without "normalization" or the
like. The worst-case performance is poor (although still linear), but
since Attrs are added by the slice, it is unlikely to arise.

We also split off the list of Attrs into a separate type, AttrList,
with an eye towards the iterator proposal. (This CL includes a copy of
the Iter[E] type, as a demonstration.)

The API changes look like this:

	type Record

	func (r *Record) Attrs() AttrList	 // repurposed
	func (r *Record) AddAttrs(attrs ...Attr) // calls AttrList.Add
	// Record.AddAttr removed
	// Record.At(int) removed

	type AttrList

	func (al *AttrList) Add(attrs ...Attr)
	func (al AttrList) Append(attrs []Attr) []Attr	// replaces the old Record.Attrs
	func (al AttrList) Len() int			// replaces Record.NumAttrs
	func (al AttrList) Range() Iter[Attr]

Thanks to Alan Donovan for extensive discussion.

Change-Id: I8b9b06c710aa7cea70c9379bee35c367bb21c1c3